feat(auth): add provider for supabase sessions
